rename FooterBtn with SocialMediaBtn and move it at the same level as footer
